more JS work, add nl_NL

+++ markasjunk2.js
@@ -20,32 +20,38 @@ function rcmail_markasjunk2(prop)
 	if (!rcmail.env.uid && (!rcmail.message_list || !rcmail.message_list.get_selection().length))
 		return;
 
+	if (!prop)
+		prop = 'markasjunk2';
+
 	var prev_sel = null;
 
 	// also select childs of (collapsed) threads
-	if (rcmail.env.uid) {
-		if (rcmail.message_list.rows[rcmail.env.uid].has_children && !rcmail.message_list.rows[rcmail.env.uid].expanded) {
-			if (!rcmail.message_list.in_selection(rcmail.env.uid)) {
-				prev_sel = rcmail.message_list.get_selection();
-				rcmail.message_list.select_row(rcmail.env.uid);
-			}
+	if (rcmail.message_list) {
+		if (rcmail.env.uid) {
+			if (rcmail.message_list.rows[rcmail.env.uid].has_children && !rcmail.message_list.rows[rcmail.env.uid].expanded) {
+				if (!rcmail.message_list.in_selection(rcmail.env.uid)) {
+					prev_sel = rcmail.message_list.get_selection();
+					rcmail.message_list.select_row(rcmail.env.uid);
+				}
 
-			rcmail.message_list.select_childs(rcmail.env.uid);
-			rcmail.env.uid = null;
+				rcmail.message_list.select_childs(rcmail.env.uid);
+				rcmail.env.uid = null;
+			}
+			else if (rcmail.message_list.get_single_selection() == rcmail.env.uid) {
+				rcmail.env.uid = null;
+			}
 		}
-		else if (!rcmail.message_list.in_selection(rcmail.env.uid)) {
-			prev_sel = rcmail.message_list.get_single_selection();
-			rcmail.message_list.remove_row(rcmail.env.uid, false);
-		}
-		else if (rcmail.message_list.get_single_selection() == rcmail.env.uid) {
-			rcmail.env.uid = null;
+		else {
+			selection = rcmail.message_list.get_selection();
+			for (var i in selection)
+				rcmail.message_list.select_childs(selection[i]);
 		}
 	}
 
 	var uids = rcmail.env.uid ? rcmail.env.uid : rcmail.message_list.get_selection().join(',');
 
 	rcmail.set_busy(true, 'loading');
-	rcmail.http_post('plugin.markasjunk2', '_uid='+uids+'&_mbox='+urlencode(rcmail.env.mailbox), true);
+	rcmail.http_post('plugin.' + prop, '_uid='+uids+'&_mbox='+urlencode(rcmail.env.mailbox), true);
 
 	if (prev_sel) {
 		rcmail.message_list.clear_selection();
@@ -57,42 +63,7 @@ function rcmail_markasjunk2(prop)
 
 function rcmail_markasnotjunk2(prop)
 {
-	if (!rcmail.env.uid && (!rcmail.message_list || !rcmail.message_list.get_selection().length))
-		return;
-
-	var prev_sel = null;
-
-	// also select childs of (collapsed) threads
-	if (rcmail.env.uid) {
-		if (rcmail.message_list.rows[rcmail.env.uid].has_children && !rcmail.message_list.rows[rcmail.env.uid].expanded) {
-			if (!rcmail.message_list.in_selection(rcmail.env.uid)) {
-				prev_sel = rcmail.message_list.get_selection();
-				rcmail.message_list.select_row(rcmail.env.uid);
-			}
-
-			rcmail.message_list.select_childs(rcmail.env.uid);
-			rcmail.env.uid = null;
-		}
-		else if (!rcmail.message_list.in_selection(rcmail.env.uid)) {
-			prev_sel = rcmail.message_list.get_single_selection();
-			rcmail.message_list.remove_row(rcmail.env.uid, false);
-		}
-		else if (rcmail.message_list.get_single_selection() == rcmail.env.uid) {
-			rcmail.env.uid = null;
-		}
-	}
-
-	var uids = rcmail.env.uid ? rcmail.env.uid : rcmail.message_list.get_selection().join(',');
-
-	rcmail.set_busy(true, 'loading');
-	rcmail.http_post('plugin.markasnotjunk2', '_uid='+uids+'&_mbox='+urlencode(rcmail.env.mailbox), true);
-
-	if (prev_sel) {
-		rcmail.message_list.clear_selection();
-
-		for (var i in prev_sel)
-			rcmail.message_list.select_row(prev_sel[i], CONTROL_KEY);
-	}
+	rcmail_markasjunk2('markasnotjunk2');
 }
 
 rcmail.rcmail_markasjunk2_move = function(mbox, uid) {
@@ -100,44 +71,30 @@ rcmail.rcmail_markasjunk2_move = function(mbox, uid) {
 	var prev_sel = null;
 	var a_uids = uid.split(",");
 
-	if (rcmail.message_list && a_uids.length == 1) {
-		if (!rcmail.message_list.in_selection(a_uids[0]))
-			prev_sel = rcmail.message_list.get_single_selection();
-
+	if (rcmail.message_list && a_uids.length == 1 && !rcmail.message_list.in_selection(a_uids[0])) {
 		rcmail.env.uid = a_uids[0];
-		rcmail.message_list.remove_row(a_uids[0], false);
+		rcmail.message_list.remove_row(rcmail.env.uid, false);
+	}
+	else if (rcmail.message_list && !rcmail.message_list.in_selection(a_uids[0])) {
+		prev_sel = rcmail.message_list.get_selection();
+		rcmail.message_list.clear_selection();
+
+		for (var i in a_uids)
+			rcmail.message_list.select_row(a_uids[i], CONTROL_KEY);
 	}
 
-	rcmail.move_messages(mbox);
+	if (mbox)
+		rcmail.move_messages(mbox);
+	else
+		rcmail.delete_messages();
 
 	rcmail.env.uid = prev_uid;
 
 	if (prev_sel) {
 		rcmail.message_list.clear_selection();
-		rcmail.message_list.select(prev_sel);
-	}
-}
 
-rcmail.rcmail_markasjunk2_delete = function(uid) {
-	var prev_uid = rcmail.env.uid;
-	var prev_sel = null;
-	var a_uids = uid.split(",");
-
-	if (rcmail.message_list && a_uids.length == 1) {
-		if (!rcmail.message_list.in_selection(a_uids[0]))
-			prev_sel = rcmail.message_list.get_single_selection();
-
-		rcmail.env.uid = a_uids[0];
-		rcmail.message_list.remove_row(a_uids[0], false);
-	}
-
-	rcmail.delete_messages();
-
-	rcmail.env.uid = prev_uid;
-
-	if (prev_sel) {
-		rcmail.message_list.clear_selection();
-		rcmail.message_list.select(prev_sel);
+		for (var i in prev_sel)
+			rcmail.message_list.select_row(prev_sel[i], CONTROL_KEY);
 	}
 }
